Remove old ChatBot.java and Replace with ChatBot package
