--- data/list.txt
@@ -1,2 +1,0 @@
-D |   | project | 2023-09-01
-E |   | birthday | 2023-09-01 12AM to 12PM
