--- src/main/java/Deadline.java
@@ -1,20 +1,0 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
-public class Deadline extends Task {
-    private LocalDate date;
-    public Deadline(String task, LocalDate date) {
-        super(task);
-        this.date = date;
-    }
-
-    @Override
-    public String toList() {
-        return "D" + super.toList() + " | " + this.date.format(DateTimeFormatter.ofPattern("yyyy-MM-dd"));
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
-    }
-}
