--- src/main/java/DukeException.java
@@ -1,11 +1,0 @@
-public class DukeException extends Exception{
-
-    public DukeException(String message) {
-        super(message);
-    }
-
-    @Override
-    public String toString() {
-        return super.getMessage();
-    }
-}
