--- src/main/java/rayshawn/chatbot/commands/HelpCommand.java
@@ -1,22 +1,0 @@
-package rayshawn.chatbot.commands;
-
-import java.security.cert.CertPath;
-
-public class HelpCommand extends Command {
-    public static final String COMMAND_WORD = "help";
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Shows program usage instructions.\n" +
-            "Example: " + COMMAND_WORD;
-
-    public CommandResult execute() {
-        return new CommandResult(ToDoCommand.MESSAGE_USAGE +
-                "\n" + DeadlineCommand.MESSAGE_USAGE +
-                "\n" + EventCommand.MESSAGE_USAGE +
-                "\n" + DeleteCommand.MESSAGE_USAGE +
-                "\n" + MarkCommand.MESSAGE_USAGE +
-                "\n" + UnmarkCommand.MESSAGE_USAGE +
-                "\n" + ListCommand.MESSAGE_USAGE +
-                "\n" + HelpCommand.MESSAGE_USAGE +
-                "\n" + ByeCommand.MESSAGE_USAGE
-        );
-    }
-}
