--- src/main/java/rayshawn/chatbot/commands/IncorrectCommand.java
@@ -1,14 +1,0 @@
-package rayshawn.chatbot.commands;
-
-public class IncorrectCommand extends Command{
-    public final String feedbackToUser;
-
-    public IncorrectCommand(String feedbackToUser) {
-        this.feedbackToUser = feedbackToUser;
-    }
-
-    @Override
-    public CommandResult execute() {
-        return new CommandResult(feedbackToUser);
-    }
-}
