+++ src/rayshawn/chatbot/messages/Messages.java
@@ -1,0 +1,14 @@
+package rayshawn.chatbot.messages;
+
+public class Messages {
+    public static final String LINE_BREAK = "------------------------------------------------------------";
+    public static final String WELCOME_MESSAGE = "Hello! I'm Desolute";
+    public static final String HELP_MESSAGE = "What can I do for you?";
+    public static final String GOODBYE_MESSAGE = "Bye. Hope to see you again soon!";
+    public static final String INIT_FAILED_MESSAGE = "Failed to initialise chat bot application. Exiting...";
+    public static final String INVALID_COMMAND_FORMAT_MESSAGE = "Incorrect command Format! \n%1$s";
+    public static final String INVALID_TASK_NUMBER_MESSAGE = "This task number is invalid!";
+    public static final String LIST_COUNT_MESSAGE = "You have %d tasks in the list.";
+    public static final String LIST_HEADER_MESSAGE = "Here are the tasks in your list:";
+
+}
