--- tasklist.txt
@@ -1,2 +1,0 @@
-T |   | list
-T |   | list
